Rename compile.rs -> compiler.rs
