+++ src/bytecode/src/compiler.rs
